working through ikine issues

+++ armSupportCode/RobotControl.h
@@ -35,12 +35,12 @@ class RobotControl {
     void  ReadMotors(dynamixel::GroupSyncRead &syncReadPacket);
     int   WriteToMotors(bool &addParamResult, dynamixel::GroupSyncWrite &syncWritePacket);
 
-    const float _A1A2,      _L1,        _L2;
-    const float _OFFSET,    _PHI,       _H_OF_L2;
-    const float _Q1_MIN,    _Q1_MAX;
-    const float _Q2_LIMIT;
-    const float _Q4_MIN,    _Q4_MAX;
-    const float _INNER_R,   _Z_LIMIT;
+    const float  _A1A2,      _L1,        _L2;
+    const double _OFFSET,    _PHI,       _H_OF_L2;
+    const double _Q1_MIN,    _Q1_MAX;
+    const double _Q2_LIMIT;
+    const double _Q4_MIN,    _Q4_MAX;
+    const double _INNER_R,   _Z_LIMIT;
     float J_M[3][3] = {{0.0f}};
     int   qPresCts_M[3],  qDotPresCts_M[3];
     float qPres_M[3],     qDotPres_M[3];
